simplify stream pipeline for concurrent decoding. rename option for mixmap-peermaps to have a hook into the feature.decoder onItem  processing. moves to doing feature detection, if we can support window.Worker, then go ahead and use the web worker setup to decode features with
