+++ lib/features.js
@@ -1,6 +1,6 @@
 var decode = require('georender-pack/decode')
-var ConcurrentWorker = require('./concurrent-worker')
-var decodeWorker = require('./feature-decode-worker')
+var ConcurrentDecoder = require('./concurrent-decoder')
+
 
 module.exports = Features
 
@@ -13,19 +13,16 @@ function Features(opts) {
   this._queryIndexIds = new Map
 
   if (!opts) opts = {}
-  if (!opts.decoder) opts.decoder = {}
-  if (typeof opts.decoder.onItem !== 'function') opts.decoder.onItem = noop
+  if (typeof opts.decoderOnItem !== 'function') opts.decoderOnItem = noop
 
-  this.decoder = ConcurrentWorker(decodeWorker, {
+  this.decoder = ConcurrentDecoder({
     count: 4,
     itemId: function (row) {
       return `${row.queryIndex}-${row.rowCount}`
     },
     onItem: function (row) {
       self.addBufferDecoded(row)
-      // TODO this is an ugly way to pass this through
-      // can we just have the parent do its own onItem processing?
-      opts.decoder.onItem()
+      opts.decoderOnItem(row)
     },
   })
 }
@@ -59,32 +56,8 @@ Features.prototype.addBufferDecoded = function ({ queryIndex, buffer, decoded })
 }
 
 Features.prototype.addBuffer = function (queryIndex, buffer) {
-  var d = decode([buffer])
-  var id = getId(d)
-  // can not yet handle multiple ids with separate geometry
-  if (this._decoded.has(id)) {
-    var buffers = this._buffers.get(id)
-    var exists = false
-    for (var i = 0; i < buffers.length; i++) {
-      if (Buffer.compare(buffer, buffers[i]) === 0) {
-        exists = true
-        break
-      }
-    }
-    if (exists === false) {
-      this._decoded.get(id).push(d)
-      buffers.push(buffer) 
-    }
-  } else {
-    this._decoded.set(id, [d])
-    this._buffers.set(id, [buffer])
-  }
-  this._count.set(id, (this._count.get(id) || 0) + 1)
-  if (this._queryIndexIds.has(queryIndex)) {
-    this._queryIndexIds.get(queryIndex).push(id)
-  } else {
-    this._queryIndexIds.set(queryIndex, [id])
-  }
+  var decoded = decode([buffer])
+  this.addBufferDecoded({ queryIndex, buffer, decoded })
 }
 
 Features.prototype._remove = function (id) {
